Update Config.py (Folder Gen and syntax fixes)
When the program is executed for the first time, it will now check for the existence of the 'CrossPixel' temporary folder. If the folder is not present, the program will automatically generate it. This new approach has been implemented to address issues encountered in the previous system. A standalone function has been created to handle this task, which helps prevent potential issues in the future.